requirements.txt
requests==2.32.3
requests-oauthlib==2.0.0
python-dotenv==1.1.0
cryptography==42.0.0